requirements.txt
# AI & Machine Learning
torch
transformers
faster-whisper
huggingface_hub
safetensors

# Audio Processing
numpy
sounddevice
pykakasi

# Utilities & Networking
yt-dlp
python-dotenv
requests